backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
langchain
langchain-community
google-generativeai
faiss-cpu
pypdf
sentence-transformers
